add status indicator during data copying.
limit memory use during raw data copying to 1gb, also increases responsiveness of status indicator and program in general.

Change behavior to print warning if number of neural channels in prm file is different than the number in the prb file. This would happen if you recorded all the channels on a bad probe, but omit bad sites from the prb file. The program (and prm) still has to acknowledge these bad channels are in the array.

Update behavior data copying behavior to copy the root node of the Voyeur h5 instead of copying children. This ensures that attributes of the root group are carried over to the new file.

Now adds attributes to the run group specifying the binary filename and the Voyeur filename for record-keeping purposes. This can help track down problems where the user is unsure if the correct files were specified during processing.